+++ druid/src/menu.rs
@@ -107,9 +107,10 @@
 
 use std::num::NonZeroU32;
 
+use crate::keyboard_types::Key;
 use crate::kurbo::Point;
-use crate::shell::{HotKey, KeyCompare, Menu as PlatformMenu, RawMods, SysMods};
-use crate::{commands, Command, Data, Env, KeyCode, LocalizedString, Selector};
+use crate::shell::{HotKey, IntoKey, Menu as PlatformMenu, RawMods, SysMods};
+use crate::{commands, Command, Data, Env, LocalizedString, Selector};
 
 /// A platform-agnostic description of an application, window, or context
 /// menu.
@@ -195,7 +196,7 @@ impl<T> MenuItem<T> {
     /// # // hide the type param in or example code by letting it be inferred here
     /// # MenuDesc::<u32>::empty().append(item);
     /// ```
-    pub fn hotkey(mut self, mods: impl Into<Option<RawMods>>, key: impl Into<KeyCompare>) -> Self {
+    pub fn hotkey(mut self, mods: impl Into<Option<RawMods>>, key: impl IntoKey) -> Self {
         self.hotkey = Some(HotKey::new(mods, key));
         self
     }
@@ -615,7 +616,7 @@ pub mod sys {
                     LocalizedString::new("win-menu-file-exit"),
                     commands::QUIT_APP,
                 )
-                .hotkey(RawMods::Alt, KeyCode::F4)
+                .hotkey(RawMods::Alt, Key::F4)
             }
         }
     }