+++ druid/src/text/text_input.rs
@@ -15,8 +15,7 @@
 //! Map input to `EditAction`s
 
 use super::Movement;
-use crate::{HotKey, KeyCode, SysMods};
-use druid_shell::{KeyEvent, KeyModifiers};
+use crate::{HotKey, KbKey, KeyEvent, Modifiers, SysMods};
 
 // This following enumerations are heavily inspired by xi-editors enumerations found at
 // https://github.com/xi-editor/xi-editor/blob/e2589974fc4050beb33af82481aa71b258358e48/rust/core-lib/src/edit_types.rs
@@ -42,7 +41,7 @@ pub enum EditAction {
 pub struct MouseAction {
     pub row: usize,
     pub column: usize,
-    pub mods: KeyModifiers,
+    pub mods: Modifiers,
 }
 
 pub trait TextInput {
@@ -68,40 +67,40 @@ impl TextInput for BasicTextInput {
             // Select all (Ctrl+A || Cmd+A)
             k_e if (HotKey::new(SysMods::Cmd, "a")).matches(k_e) => EditAction::SelectAll,
             // Jump left (Ctrl+ArrowLeft || Cmd+ArrowLeft)
-            k_e if (HotKey::new(SysMods::Cmd, KeyCode::ArrowLeft)).matches(k_e)
-                || HotKey::new(None, KeyCode::Home).matches(k_e) =>
+            k_e if (HotKey::new(SysMods::Cmd, KbKey::ArrowLeft)).matches(k_e)
+                || HotKey::new(None, KbKey::Home).matches(k_e) =>
             {
                 EditAction::Move(Movement::LeftOfLine)
             }
             // Jump right (Ctrl+ArrowRight || Cmd+ArrowRight)
-            k_e if (HotKey::new(SysMods::Cmd, KeyCode::ArrowRight)).matches(k_e)
-                || HotKey::new(None, KeyCode::End).matches(k_e) =>
+            k_e if (HotKey::new(SysMods::Cmd, KbKey::ArrowRight)).matches(k_e)
+                || HotKey::new(None, KbKey::End).matches(k_e) =>
             {
                 EditAction::Move(Movement::RightOfLine)
             }
             // Select left (Shift+ArrowLeft)
-            k_e if (HotKey::new(SysMods::Shift, KeyCode::ArrowLeft)).matches(k_e) => {
+            k_e if (HotKey::new(SysMods::Shift, KbKey::ArrowLeft)).matches(k_e) => {
                 EditAction::ModifySelection(Movement::Left)
             }
             // Select right (Shift+ArrowRight)
-            k_e if (HotKey::new(SysMods::Shift, KeyCode::ArrowRight)).matches(k_e) => {
+            k_e if (HotKey::new(SysMods::Shift, KbKey::ArrowRight)).matches(k_e) => {
                 EditAction::ModifySelection(Movement::Right)
             }
             // Move left (ArrowLeft)
-            k_e if (HotKey::new(None, KeyCode::ArrowLeft)).matches(k_e) => {
+            k_e if (HotKey::new(None, KbKey::ArrowLeft)).matches(k_e) => {
                 EditAction::Move(Movement::Left)
             }
             // Move right (ArrowRight)
-            k_e if (HotKey::new(None, KeyCode::ArrowRight)).matches(k_e) => {
+            k_e if (HotKey::new(None, KbKey::ArrowRight)).matches(k_e) => {
                 EditAction::Move(Movement::Right)
             }
             // Backspace
-            k_e if (HotKey::new(None, KeyCode::Backspace)).matches(k_e) => EditAction::Backspace,
+            k_e if (HotKey::new(None, KbKey::Backspace)).matches(k_e) => EditAction::Backspace,
             // Delete
-            k_e if (HotKey::new(None, KeyCode::Delete)).matches(k_e) => EditAction::Delete,
+            k_e if (HotKey::new(None, KbKey::Delete)).matches(k_e) => EditAction::Delete,
             // Actual typing
-            k_e if k_e.key_code.is_printable() => {
-                if let Some(chars) = k_e.text() {
+            k_e if key_event_is_printable(k_e) => {
+                if let KbKey::Character(chars) = &k_e.key {
                     EditAction::Insert(chars.to_owned())
                 } else {
                     return None;
@@ -111,5 +110,24 @@ impl TextInput for BasicTextInput {
         };
 
         Some(action)
+    }
+}
+
+/// Determine whether a keyboard event contains insertable text.
+fn key_event_is_printable(event: &KeyEvent) -> bool {
+    if let KbKey::Character(_) = &event.key {
+        if event.mods.ctrl() || event.mods.meta() {
+            return false;
+        }
+        // On mac, Alt functions more like AltGr.
+        #[cfg(not(target_os = "macos"))]
+        {
+            if event.mods.alt() {
+                return false;
+            }
+        }
+        true
+    } else {
+        false
     }
 }